+++ cpp/oneapi/dal/backend/primitives/BUILD
@@ -37,6 +37,7 @@ dal_collect_modules(
         "lapack",
         "optimizers",
         "objective_function",
+        "frontier",
         "placement",
         "reduction",
         "regression",
@@ -88,6 +89,7 @@ dal_collect_test_suites(
         "lapack",
         "optimizers",
         "objective_function",
+        "frontier",
         "placement",
         "reduction",
         "regression",
