+++ cpp/oneapi/dal/backend/primitives/frontier.hpp
@@ -1,3 +1,3 @@
-#include "oneapi/dal/backend/primitives/frontier/frontier_dpc.hpp"
+#include "oneapi/dal/backend/primitives/frontier/frontier.hpp"
 #include "oneapi/dal/backend/primitives/frontier/advance_dpc.hpp"
 #include "oneapi/dal/backend/primitives/frontier/graph.hpp"
